Made create/edit a dynamic decision
The agent is asked to review the pr list for the provided
source and base branches.  If there isn't a PR on the list,
it will create a new one.  If there is one PR, it will edit
that one.  If there are multiple (rare occurance), it should
ask for use input.